re-organize files to right folders
