+++ release/pipeline/operators/gcs_to_gcs.py
